fixed question arrays
